Order Pipeline Listener Weekly Payment Listener Debugging and Code Cleaup

+++ app/src/main/java/com/example/kafiesta/utilities/dialog/ProductEditDialog.kt
@@ -122,7 +122,7 @@ class ProductEditDialog(
                         name = pName,
                         description = pDesc,
                         imageURL = product.imageURL,
-                        price = pPrice.toDouble(),
+                        price = pPrice,
                         tags = pTags,
                         status = pStatus
                     )
